converter/requirements.txt
requests>=2.28.0
Pillow>=9.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
